+++ examples/app.rs
@@ -1,18 +1,16 @@
 // Import dependencies from Feather
-use feather::HttpResponse;
-use feather::middleware::{Logger, MiddlewareResult, Cors};
-use feather::{App, AppConfig};
+use feather::{App,MiddlewareResult};
+use feather::middleware::Logger;
+use feather::{Request, Response};
+
 // Main function - no async here!
 fn main() {
-    // Create instance of AppConfig with 4 threads
-    let config = AppConfig { threads: 32 };
-
     // Create a new instance of App
-    let mut app = App::new(config);
-    app.use_middleware(Cors::default());
+    let mut app = App::new();
+    
     // Define a route for the root path
-    app.get("/", |_request: &mut _, response: &mut _| {
-        *response = HttpResponse::ok("Hello from Feather!");
+    app.get("/", |_request: &mut Request, response: &mut Response| {
+        response.send_text("Hello, world!");
         MiddlewareResult::Next
     });
     // Use the Logger middleware for all routes