+++ examples/middleware.rs
@@ -1,6 +1,5 @@
 // Import dependencies from Feather
-
-use feather::{App, AppConfig, HttpRequest,HttpResponse};
+use feather::{App, AppConfig, Request, Response};
 // Import the Middleware trait and some common middleware primitives
 use feather::middleware::{Logger, Middleware, MiddlewareResult};
 
@@ -11,7 +10,7 @@ struct Custom;
 // The Middleware trait defines a single method `handle`,
 // which can mutate the request and response objects, then return a `MiddlewareResult`.
 impl Middleware for Custom {
-    fn handle(&self, request: &mut HttpRequest, _response: &mut HttpResponse) -> MiddlewareResult {
+    fn handle(&self, request: &mut Request, _response: &mut Response) -> MiddlewareResult {
         // Do stuff here
         println!("Now running some custom middleware (struct Custom)!");
         println!("And there's a request with path: {:?}", request.uri);
@@ -21,9 +20,9 @@ impl Middleware for Custom {
 }
 
 fn main() {
-    // Define an app
-    let config = AppConfig { threads: 4 };
-    let mut app = App::new(config);
+    // Define an app with config this time
+    let config = AppConfig { threads: 8 }; // Curently Config only sets the number of threads
+    let mut app = App::with_config(config);
 
     // Use the builtin Logger middleware for all routes
     app.use_middleware(Logger);
@@ -32,14 +31,14 @@ fn main() {
     app.use_middleware(Custom);
 
     // Use another middleware defined by a function for all routes
-    app.use_middleware(|_request: &mut HttpRequest, _response: &mut HttpResponse| {
+    app.use_middleware(|_request: &mut Request, _response: &mut Response| {
         println!("Now running some custom middleware (closure)!");
         MiddlewareResult::Next
     });
 
     // Define a route
-    app.get("/", |_request: &mut _, response: &mut _| {
-        *response = HttpResponse::ok("Hello from Feather!");
+    app.get("/", |_request: &mut Request, response: &mut Response| {
+        response.send_text("Hello, world!");
         MiddlewareResult::Next
     });